pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["znammikhail <znamenskymihail@gmail.com>"]
packages = [
  { include = "gendiff" },
]
repository = 'https://github.com/znammikhail/python-project-50'

[tool.poetry.dependencies]
python = "^3.8.1"
PyYAML = "^6.0"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
